+++ packages/shared/src/molecules/MultiMonthPicker/index.stories.tsx
@@ -1,11 +1,11 @@
 import { Meta } from '@storybook/react'
 import { useForm } from 'react-hook-form'
 import React from 'react'
-import MultiDateInput from './index'
+import MultiMonthPicker from './index'
 
-const config: Meta<typeof MultiDateInput> = {
-  title: 'MultiDateInput',
-  component: MultiDateInput,
+const config: Meta<typeof MultiMonthPicker> = {
+  title: 'MultiMonthPicker',
+  component: MultiMonthPicker,
 }
 
 export default config
@@ -29,7 +29,7 @@ export const Primary = () => {
 
   return (
     <form onSubmit={onSubmit}>
-      <MultiDateInput
+      <MultiMonthPicker
         startDateRegister={register('startDate', {
           required: { value: true, message: '필수 값입니다' },
         })}
