+++ packages/shared/src/molecules/index.ts
@@ -4,7 +4,7 @@ export { default as Select } from './Select'
 export { default as SearchInput } from './SearchInput'
 export { default as Header } from './Header'
 export { default as MultiRangeSlider } from './MultiRangeSlider'
-export { default as MultiDateInput } from './MultiDateInput'
+export { default as MultiMonthPicker } from './MultiMonthPicker'
 export { default as FourImageInputs } from './FourImageInputs'
 export { default as DoubleInput } from './DoubleInput'
